+++ app/expenses/(list)/expense-row-actions.tsx
@@ -5,6 +5,7 @@ import { Button } from '@/components/ui/button';
 import { Dialog, DialogTrigger } from '@/components/ui/dialog';
 import { TableCell } from '@/components/ui/table';
 import axios from 'axios';
+import Link from 'next/link';
 import { useRouter } from 'next/navigation';
 import { toast } from 'sonner';
 
@@ -15,28 +16,37 @@ const ExpenseRowActions = ({ id }: { id: string }) => {
     try {
       await axios.delete(`/api/expenses/${id}`);
       router.refresh();
+
       toast.success('Expense deleted successfully');
     } catch (error) {
       console.error(error);
+
       toast.error('Failed to delete expense');
     }
   };
 
   return (
-    <TableCell>
-      <Dialog>
-        <DialogTrigger asChild>
-          <Button variant="destructive">Delete</Button>
-        </DialogTrigger>
-        <ConfirmationDialog
-          title="Are you sure you want to delete this expense?"
-          description="This operation is irreversible."
-          onConfirm={() => handleDelete(id)}
-          confirmButtonText="Delete"
-          cancelButtonText="Cancel"
-        />
-      </Dialog>
-    </TableCell>
+    <>
+      <TableCell>
+        <Dialog>
+          <DialogTrigger asChild>
+            <Button variant="destructive">Delete</Button>
+          </DialogTrigger>
+          <ConfirmationDialog
+            title="Are you sure you want to delete this expense?"
+            description="This operation is irreversible."
+            onConfirm={() => handleDelete(id)}
+            confirmButtonText="Delete"
+            cancelButtonText="Cancel"
+          />
+        </Dialog>
+      </TableCell>
+      <TableCell>
+        <Button variant="outline" asChild>
+          <Link href={`/expenses/edit/${id}`}>Edit</Link>
+        </Button>
+      </TableCell>
+    </>
   );
 };
