+++ app/expenses/edit/[id]/loading.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const EditExpenseLoadingPage = () => {
+  return <div>EditExpenseLoadingPage</div>;
+};
+
+export default EditExpenseLoadingPage;
